Add driver for Windows Surface Integration Device
Implements performance-mode setting via the EC, as can be done through
the performance-slider on Windows. This affects at least the behavior of
fans, but may also influence other aspects of the device. Performance
modes can be set via the `perf_mode` sysfs attribute. The respective
numeric values that can be set are (with their respective name on
Windows):

  Value  Name
      1  Recommended (Default)
      2  Battery Saver
      3  Better Performance
      4  Best Performance

The performance mode can also be set when the module is loaded using the
module parameter `perf_mode_init` and when the module is unloaded using
the parameter `perf_mode_exit`, where the value of 0 specifies that the
mode is kept as-is and not modified (this is the default behavior).