Enforce dGPU power after Root Port power-down
It seems that in some instances (initial power-down) the direct dGPU
power does not get cut. Make sure that the power is cut by explicitly
calling shps_dgpu_dsm_set_power(). Note that this method checks if the
dGPU is being supplied with power, making it a no-op in case powering
down the Root Port worked as intended.