Load PCIe Root Port address from ACPI
Load the PCIe Root Port of the dGPU from the _DSM. If the dGPU is
disabled or physically disconnected, searching for it and its Root Port
by its vendor and device ID will not work. Instead, we can leverage
ACPI, which provides PCIe addresses for both the Root Port (as RP5_PCIE)
as well as the dGPU itself (DGPU_PCIE).