Implement no-ACK type frame
Implement type 0x00 control frame. This type of frame is followed by a
command data frame and optional payload, but does not require an ACK in
return. The value 0x00 is only used in the control frame, whereas the
command data frame still has type 0x80.

This frame-type seems to have been introduced on the Surface Laptop 3
and is used for highly frequent events. Currently known events that use
this type are keyboard and touchpad input events. The introduction of
this type is likely a direct result of routing the touchpad through SAM,
for which waiting for ACKs before sending a new event may not be
feasible due to the rate with which events are being produced.