Check event delay function before calling it
If no event handler has been set or an event handler has been removed,
the delay function will be NULL. Nevertheless, events may still be sent
by the EC (whether configured to do so or not), so we need to check the
delay function and ensure it is non-NULL before calling it.

In case it is NULL, we assume a delay of zero, forwarding it to the
workq-queue and let the work function take care of logging it as
unhandled event.

Reported by: Blaž Hrastnik <blaz@mxxn.io>