Replace DMA hack by late_initcall
To ensure a stable UART/tty communication with the EC, DMA should be
used. Previously, a hack relying on the internals of the tty/serdev
device was used to do that. This hack does not work on the Surface
Laptop 3 AMD (15") model. Thus replace the hack with a late_initcall of
the module, which should have a similar effect.