Implement wakeup IRQ for SAM over SSH
The SAM EC can wake up the system via the GPIO interrupt defined in ACPI
on the SSH device. This interrupt only seems to be triggered when the EC
is suspended via the corresponding SSH request. Events triggering this
interrupt include the battery trip point and, on the Surface Book 2,
detaching and attaching the clipboard (while suspended).

Also add an interrupt routine printing occurences of the interrupt for
debugging.