Synchronize to target power-state on resume
During resume, the PCIe core drivers will turn on the dGPU if it has
been turned off by resuming the root port. The on-state consistency will
be insured by the SAN RQSG handler, however, using that to turn-off the
dGPU causes some synchrinization issues. Thus we have to do that in
pm_complete for now.