Wake device when base is being detached/attached
When attaching the device while the dGPU is still powered up (i.e. the
device has not been resumed after detaching to correct for this), the EC
will power up, but the device won't. Fix this by triggering a wake-ups
via the base-presence GPIO.

This resembles the behavior on Windwos, however, on Windows the wake-ups
are triggered by the EC via its GPIO directly instead of relying on the
base-presence GPIO. A result of this is that on Windows they are
triggered earlier, i.e. before actually physically disconnecting the
device parts. As these EC wake-ups are currently disabled on Linux by
default due to spurious wake-ups, we ensure that this somewhat critical
wakeup still happens by relying on that secondary GPIO pin.