Lock/unlock the DTX latch when dGPU is powered on
Lock/unlock the DTX latch when the dGPU is powered on. This prevents the
base from being detached (unless the EC receives a commence signal from
the DTX deamon). The main impact of this is the base not detaching when
the device is suspended and the dGPU is turned on.