Implement vga-switcheroo handler and dummy client
Implement a vga-switcheroo handlerto turn on/off the dGPU to allow for
basic vga-switcheroo support. Also implement a dummy-client to allow
this to work with the proprietary nvidia driver.

This does not seem to work with the nouveau driver for (seemingly)
unrelated issues and has been tested with the proprietary nvidia driver
and bumblebee/optirun/pvkrun. Some configuration and the correct module
parameters are required.