Load protocol multi-byte values as unaligned
Explicitly load protocol multi-byte values as unaligned. The multi-byte
values inside the protocol structs are not aligned. Furthermore, we
cannot rely on alignment as they are casted from a raw input buffer.

While we currently only expect the driver to load on x86 where this is
not an issue, make it explicit nontheless, if for nothing else then to
document these facts.