Termination bytes are actually CRC
The previously described termination bytes [0xff, 0xff] are actually the
CRC (ccitt-false initialized with 0xffff) of a zero-length sequence.
This means that if the payload is not present, i.e. has length zero, the
CRC of this payload is 0xffff, or [0xff, 0xff]. In full, this means that
the CRC of the payload is always present and the "termination byte
sequence" actually does not exist.