Refactor ptx to use reference counting
Employ reference counting for packets. This simplifies lifetime
managment for packets and allows us to provide a helpful guarantee for
the completion callback of sequenced packets: If a sequenced packet is
being completed successfully, the callback will run on the receiver
thread (i.e. the thread calling ssh_ptx_acknowledge). This allows
synchronization with regards other received messages, something that was
not guaranteed before.