Hook new implementation into old API
This commit replaces the old synchronous implementation with the new
asynchonous one under the hood, while still keeping the old API. This
way we can test most parts of the new implementation without having to
implement and adapt to a new API. Note that the new implementation is
inherently asynchronous under the hood and only synchronized in the
top-level request function.

The new API will be implemented later, including support for completely
asynchronous requests.