Fix wakeup bug for transmitter thread
Force-wakeup transmitter thread if packets have been re-submitted.
Re-submitted packets are allowed to be transmitted even if they're
already pending. Without this, the transmitter wakeup function only
checks for the pending count, thus the thread will not be woken up if
the limit of pending packets has been reached, even if already pending
packets that have been scheduled for re-transmission could be
transmitted.