Remove state-checks for nf-registry functions
The controller state for the notifier_register/notifier_unregister
functions only needs to be "initialized". Subsequent calls to SSAM
functions (event_enable/event_disable) check their required state
themselves.

The caller now has to guarantee that the state is at least
"initialized". Note that currently, out-of-module references to a valid
struct ssam_controller reference can only be obtained by binding the
controller to a client device, thus the controller is always guaranteed
to be initialized as long as the respective client device driver is
attached.