Reduce cmpxchg calls in timeout_reaper_mod
Reduces the number of cmpxchg calls in the timeout_reaper_mod functions.
The previous implementation always forced a superfluous final cmplxchg
that was guaranteed to fail after the previous cmpxchg call has already
succeeded. This call essentially boiled down to a final read of the
value, required to trigger the loop exit condition.

Change the cmpxchg loop condition to remove that final call.