cdev: Fix OOPS when driver is unbound while files are still open
Open file references can still linger around after the driver has been
unloaded/the miscdevice deregistered. This means, that IOCTLs can still
be executed, even though the driver has already been unloaded, which can
ultimately lead to an OOPS due to access-after-free on the main struct.

Prevent this from happening by reference-counting the main struct.
Further, add explicit locking around the controller and remove its
refernce when the driver unbinds. This is required as the controller is
only guaranteed to be valid for as long as the driver is bound.