Allow registring notifiers as observers
Allow registring notifiers as observers, i.e. without them being tied to
enabling/disabling of events. This can be useful for debugging purposes,
when we do not want to change the event state on the SAM device.

Due to this, observer notifiers also only need to be associated with a
specific target category, other event properties can be skipped. Due to
this, we can now register a single notifier for all event categories.