Allow enabling and disabling of events without notifiers
Observer-notifiers can be registered for a target category, without
enabling any events or modifying the reference count thereof. The
counter-part to this is enabling events without notifiers, which
finalizes decoupling of events and notifiers.

Combined, these two parts allow us to register a single notifier per
target category, capturing all events of that category. Due to this we
can then use a single notifier callback function, whereas this could
previoulsy, with multiple registered notifiers, lead to duplicate
events. This can then be used in the surface_aggregator_cdev interface
to provide user-space with a single stream of (non-duplicate) events.