registry: Use KIP hub for Surface Book 3 base devices
It turns out that the Surface Book 3 manages the devices contained in
its detachable base via the KIP hub as well, similarly to the Surface
Pro 8 and Surface Pro X. So move them over to the KIP hub.

Right now, we (mis-)use the detachment subsystem (DTX), which is
designed for handling detachment requests and physical locking of the
base, to properly remove and re-attach Surface System Aggregator Module
(SSAM) client devices contained in the base. This system does not seem
to be intended for managing the (sub-)devices contained in the base,
which may need some time to be set up properly.

The KIP subsystem seems to be the intended subsystem for managing those
devices, thus let's use that one instead.

Note that this also changes the way in which devices on the Surface Book
3 are removed when they have been detached, specifically from normal
removal to hot-removal (avoiding further communication with the embedded
controller). It seems that the "communication timeout after device
removal" issue does also occur on the Surface Book 3, but has so far
been missed as it does not happen reliably every time. Switching to
hot-removal fixes this issue and should not have any noticable
drawbacks.

Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>