tabletsw: Add support for book mode
Devices with a type-cover (using either KIP or POS subsystems for device
posture information) have an additional "book" mode. This is currently
unsupported, leading to the warning

  surface_aggregator_tablet_mode_switch 01:03:01:00:01: unknown KIP cover state: 6

or

  surface_aggregator_tablet_mode_switch 01:26:01:00:01: unknown device posture for type-cover: 6

depending on the exact subsystem used by the device. Therefore, add
support for this state.

The book mode is triggered by the device being in portrait orientation
while the type-cover is attached and in an open position. Other than
being connected and not fully closed towards the screen, the exact state
of the type-cover does not matter (i.e. there are no sub-states for
canvas- or folded-back positions). When in this state, the type-cover is
turned off (i.e., backlight is off and no input is possible), therefore
map it to enable tablet-mode.

Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>